[만두] 6,7단계 구현

+++ src/main/java/kr/codesqaud/cafe/service/PostService.java
@@ -2,71 +2,89 @@ package kr.codesqaud.cafe.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import kr.codesqaud.cafe.domain.Member;
 import kr.codesqaud.cafe.domain.Post;
 import kr.codesqaud.cafe.dto.post.PostModifyRequest;
 import kr.codesqaud.cafe.dto.post.PostResponse;
 import kr.codesqaud.cafe.dto.post.PostWriteRequest;
-import kr.codesqaud.cafe.exception.common.Unauthorized;
+import kr.codesqaud.cafe.exception.common.UnauthorizedException;
 import kr.codesqaud.cafe.exception.post.PostNotFoundException;
-import kr.codesqaud.cafe.repository.member.MemberRepository;
+import kr.codesqaud.cafe.repository.comment.CommentRepository;
 import kr.codesqaud.cafe.repository.post.PostRepository;
-import kr.codesqaud.cafe.config.session.AccountSession;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
 public class PostService {
 
     private final PostRepository postRepository;
-    private final MemberRepository memberRepository;
+    private final CommentRepository commentRepository;
 
-    public PostService(PostRepository postRepository, MemberRepository memberRepository) {
+    public PostService(PostRepository postRepository, CommentRepository commentRepository) {
         this.postRepository = postRepository;
-        this.memberRepository = memberRepository;
+        this.commentRepository = commentRepository;
     }
 
-    public Long write(PostWriteRequest postWriteRequest, Long id) {
-        postWriteRequest.setWriterId(id);
-        return postRepository.save(postWriteRequest.toPost());
+    @Transactional
+    public Long write(PostWriteRequest postWriteRequest) {
+        return postRepository.save(postWriteRequest.toPost(Member.builder()
+            .id(postWriteRequest.getWriterId())
+            .build()));
     }
 
+    @Transactional
     public PostResponse findById(Long id) {
+        postRepository.increaseViews(id);
+        return PostResponse.from(postRepository.findById(id).orElseThrow(PostNotFoundException::new));
+    }
+
+    @Transactional(readOnly = true)
+    public PostModifyRequest findPostModifyById(Long id, Long accountSessionId) {
+        Post post = validateUnauthorized(id, accountSessionId);
+        return new PostModifyRequest(post.getId(), post.getTitle(), post.getContent());
+    }
+
+    private Post validateUnauthorized(Long id, Long accountSessionId) {
         Post post = postRepository.findById(id)
-            .orElseThrow(PostNotFoundException::new)
-            .increaseViews();
-        postRepository.increaseViews(post);
-        return PostResponse.of(post);
+            .orElseThrow(PostNotFoundException::new);
+
+        if (!post.equalsWriterId(accountSessionId)) {
+            throw new UnauthorizedException();
+        }
+
+        return post;
     }
 
     @Transactional(readOnly = true)
     public List<PostResponse> findAll() {
         return postRepository.findAll()
             .stream()
-            .map(PostResponse::of)
+            .map(PostResponse::from)
             .collect(Collectors.toUnmodifiableList());
     }
 
-    public void modify(PostModifyRequest postModifyRequest, Long id) {
-        postModifyRequest.setId(id);
-        postRepository.findById(postModifyRequest.getId())
-            .orElseThrow(PostNotFoundException::new);
+    @Transactional
+    public void modify(PostModifyRequest postModifyRequest, Long accountSessionId) {
+        validateUnauthorized(postModifyRequest.getId(), accountSessionId);
         postRepository.update(postModifyRequest.toPost());
     }
 
-    @Transactional(readOnly = true)
-    public void validateUnauthorized(Long id, AccountSession accountSession) {
-        Post findPost = postRepository.findById(id)
-            .orElseThrow(PostNotFoundException::new);
-
-        if (!findPost.equalsWriterId(accountSession.getId())) {
-            throw new Unauthorized();
-        }
+    @Transactional
+    public void delete(Long id, Long accountSessionId) {
+        validateDeleteUnauthorized(id, accountSessionId);
+        postRepository.delete(id);
+        commentRepository.deleteAllByPostId(id);
     }
 
-    public void delete(Long id) {
-        postRepository.findById(id)
-            .orElseThrow(PostNotFoundException::new);
-        postRepository.delete(id);
+    private void validateDeleteUnauthorized(Long id, Long accountSessionId) {
+        validateUnauthorized(id, accountSessionId);
+
+        boolean isNotSameWriter = commentRepository.findAllByPostId(id)
+            .stream()
+            .anyMatch(comment -> !comment.isSameWriterId(accountSessionId));
+
+        if (isNotSameWriter) {
+            throw new UnauthorizedException("게시글 작성자와 댓글 작성자가 다릅니다.");
+        }
     }
 }
